avoid conditional on critical path w/o self-modification
When displaying the source code of an existing function, the :function
command will respect settings such as 'listchars' which makes the
recompilation code fragile. (I could research all the settings that
affect the output in such a way it can no longer be :execute’d and add
code to save, reset and restore them, but even if the list of settings
is comprehensive now, more such settings may get added to Vim in the
future. Therefore it is preferrable to circumstep the issue entirely.)