Fix error for undefined variable
I think the error described in 9278c29, happening when reloading buffers
that were deleted when highlighting was enabled, and reloaded when
disabled, is due to the fact that buffer autocommands are not deleted
with the buffer, while variables are, so the autocommand looks for
variables that don't exist anymore. This fixes it.