improve terminal color approximation using DIN99
Also, recompute all highlights in syntax/colornames/*.vim

This resolves #78.

Improving the visual accuracy of approximated colours is hampered by
the fact that VimL is dog slow. The following matrix of choices results:

 Estimator function                     Performance   Visual accuracy
-------------------------------------- ------------- ----------------------
 Euclidian distance in RGB              fast          awful
 Euclidian distance in CIELAB (CIE76)   slow          bad
 Euclidian distance in DIN99            slow          reasonable
 CIE94                                  very slow     reasonable
 CIEDE2000                              intolerable   high (but imperfect)

One might hope that keeping to plain Euclidian distance but inserting
a conversion to a more uniform colour space in front of it, AKA CIE76,
would avoid too high a performance penalty while noticeably improving
accuracy. Alas, it offers the opposite on both counts: the disappointing
accuracy is not worth the disappointing performance.

CIE94 delivers accuracy that for the first time can even be considered
acceptable at all for many common colours, such as various combinations
of 33/66/99 components. Unfortunately the amount of computation required
during comparisons makes this function painfully slow. It is not too bad
in typical use with just a few colours in the viewport, but can easily
become a perceptible slowdown, and makes opening the torture test CSS
file a multi-second proposition.

The DIN99 colour space is derived from CIELAB through a transformation
that is supposed to bake in the corrections applied to distances in
CIELAB in CIE94, such that Euclidian distance in DIN99 should be largely
equivalent to CIE94. And indeed it does produce results not too far off
from CIE94, even sometimes better, while being far and away faster.
Performance is (slightly) worse yet than CIE76, yet accuracy is improved
so much that the cost in performance is worthwhile.