queue :syn and :hi commands during parse
This means only a single :exe is run per frame for all highlighting-
related commands, which *drastically* speeds up processing for frames
with a very large number of newly encountered colors in the viewport.
The magnitude of improvement is enough to make opening or paging through
the torture test CSS file nearly tolerable (or maybe better put, reduces
lag to almost purely the sluggishness of Vim’s own repaint).